rooms messages complete and cleaned up code

+++ client/src/components/Rooms.vue
@@ -35,10 +35,8 @@
 
       <v-layout align-center justify-center> <!-- this centers the contents -->
         <v-flex id="box2" class="text-xs-center" mt-2 v-if="confirm" v-model="confirm">
-            <h2>User Created <i class="material-icons">check_circle</i></h2>
-            Username:  {{this.savedUser}}
-            <br>Employee Type:  {{this.savedType}}
-            <br>Temporary Password: {{this.savedPass}}
+            <h2>Room Created <i class="material-icons">check_circle</i></h2>
+            Room Name:  {{this.savedRoomAdd}}
         </v-flex>
       </v-layout>
   
@@ -56,16 +54,22 @@
       </v-btn>
       
       <v-flex id="errorMessage" class="text-xs-center" mt-3 v-if="deleteError" v-model="deleteError">
-        {{ "Usernames not found" }}
+        {{ "This room was not found" }}
         <i class="material-icons">error</i>
       </v-flex>
+
+        <v-flex id="errorMessage" class="text-xs-center" mt-3 v-if="deleteLengthError" v-model="deleteLengthError">
+          {{ "Room Name should be between 3 and 10 characters" }}
+          <i class="material-icons">error</i>
+        </v-flex>
+
     </v-flex>
   <br>
       <v-layout align-center justify-center> <!-- this centers the contents -->
       
         <v-flex id="boxDelete" class="text-xs-center" mt-2 v-if="deleteConfirm" v-model="deleteConfirm">
-            <h2>User Deleted</h2>
-            Username:  {{this.savedDeleteName}}
+            <h2>Room Deleted</h2>
+            Room Name:  {{this.savedDeleteRoom}}
         </v-flex>
       </v-layout>
 
@@ -95,8 +99,6 @@
             </template>
         </v-data-table>
     </v-flex>
-
-    <v-text name="employees" type="text" id="employees" label="employees"  v-model="employees"/>
     
      </table>
 
@@ -114,23 +116,13 @@ export default {
       roomNameField: "", 
       deleteRoom: "",
       savedDeleteRoom: "",
-
-
-
-
-
-   //   employeeType: "", 
+      savedRoomAdd: "",
+      deleteConfirm: false,
+      confirm: false,
       lengthError: false,
       usedError: false,
-      confirm: false,
-      password: "red",
-      savedUser: "",
-      savedType: "",
-      savedPass: "",
       deleteError: false,
-      deleteName: "",
-      deleteConfirm: false,
-
+      deleteLengthError: false,
 
       // table columns
       headers: [
@@ -140,36 +132,36 @@ export default {
           sortable: true,
           value: 'roomName'
         }
-       
       ],
       rooms: {} // this items is for the datatable
-
     };
   },
+
   created() {
       this.load();
   },
 
   methods: {
 
-
-
     async load(){
-
         var roomResponse = await ApiFunctions.getRoomList();
         var parsedData = JSON.parse(roomResponse.data);
         this.rooms = parsedData.values;
     },
-    async submit() {
 
- 
-      // reset messages
+    async resetMessages(){
       this.lengthError = false
       this.usedError = false
       this.confirm = false
       this.deleteError = false
       this.deleteConfirm = false
+      this.deleteLengthError = false
+    },
 
+    async submit() {
+
+      this.resetMessages();
+      
       if (this.roomNameField.length < 3 || this.roomNameField.length > 10) {
         this.lengthError = true
       }
@@ -180,10 +172,10 @@ export default {
             roomName: this.roomNameField
           })
           this.load();
-
+          this.savedRoomAdd = this.roomNameField;
+          this.confirm = addResponse.data;
+          this.usedError = !addResponse.data; // false add changed to true error
         }
-
-
         catch (error) {
           console.log("catch condition 1")
         } 
@@ -192,32 +184,31 @@ export default {
     },
     async deleteBtn() {
 
-      // reset messages
-      this.lengthError = false
-      this.usedError = false
-      this.confirm = false
-      this.typeError = false
-      this.deleteError = false
-      this.deleteConfirm = false
-      try {
-        console.log("sending to ApiFunctions: ", this.deleteRoom)
-          const deletedConfirm = await ApiFunctions.deleteRoom({
-          roomIn: this.deleteRoom
-        })
-          this.savedDeleteRoom = this.deleteName
-          this.load();
-       //   this.deleteConfirm = true
-        
-        if (this.deletedConfirm = false) {
-          this.deleteError = true;
-        }
-      }
-      catch (error) {
-        console.log("catch condition 1")
-      } 
+      this.resetMessages();
 
-    },
-    
+      if (this.deleteRoom.length < 3 || this.deleteRoom.length > 10) {
+        this.deleteLengthError = true;
+      }
+      else {
+
+        try {
+            const deletedConfirm = await ApiFunctions.deleteRoom({
+            roomIn: this.deleteRoom
+          })
+            this.deleteError = !deletedConfirm.data
+            this.savedDeleteRoom = this.deleteRoom
+            this.load();
+
+          if (this.deleteError == false){
+            this.deleteConfirm = true;
+          }
+          
+        }
+        catch (error) {
+          console.log("catch condition 1")
+        } 
+      }
+    }
   }
 };
 </script>
